Moved fileOpener function from openbutton component to utils
