+++ packages/ketcher-react/src/script/ui/utils/index.ts
@@ -63,3 +63,5 @@ export function filterFGLib(lib, filter) {
     }, {})
   )(lib)
 }
+
+export { fileOpener } from './fileOpener'
